pyproject.toml
[tool.poetry]
name = "db2obsidian"
version = "0.1.0"
description = ""
authors = ["György Kiss <gyorgy@duck.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
db2obsidian = 'db2obsidian.cli:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
